refactor: organise catalogue
