+++ dotcom/src/routes/(main)/works/bobbin/catalogue/Crescent.svelte
@@ -1,7 +1,7 @@
 <script>
-	import Button from "../../../../lib/Button.svelte";
-	import SVG from "./SVG.svelte";
-	import { add_pattern, current } from "./store";
+	import Button from "$lib/Button.svelte";
+	import SVG from "../SVG.svelte";
+	import { add_pattern, current } from "../store";
 
 	let radius = 6;
