+++ dotcom/src/routes/(main)/works/bobbin/catalogue/Loop.svelte
@@ -1,8 +1,8 @@
 <script>
-	import Button from "../../../../lib/Button.svelte";
-	import SVG from "./SVG.svelte";
-	import { add_pattern, current } from "./store";
-	import { loop } from "./weaving.js";
+	import Button from "$lib/Button.svelte";
+	import SVG from "../SVG.svelte";
+	import { add_pattern, current } from "../store";
+	import { loop } from "../weaving.js";
 
 	let count = 2;
 	let bulge = 3 / 2;
