+++ dotcom/src/routes/(main)/works/bobbin/catalogue/Polygon.svelte
@@ -1,8 +1,8 @@
 <script>
-	import Button from "../../../../lib/Button.svelte";
-	import SVG from "./SVG.svelte";
-	import { add_pattern, current } from "./store";
-	import { on_circle } from "./weaving.js";
+	import Button from "$lib/Button.svelte";
+	import SVG from "../SVG.svelte";
+	import { add_pattern, current } from "../store";
+	import { on_circle } from "../weaving.js";
 
 	let radius = 12;
 	let sides = 5;
